requirements.txt
flask==3.1.0
pyyaml==6.0.2
docker==7.1.0
gunicorn==23.0.0
